split user invites to own controller
